Implemented new view in Out Team with categories
